% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapleyValue-package.R
\docType{package}
\name{ShapleyValue-package}
\alias{ShapleyValue}
\alias{ShapleyValue-package}
\title{ShapleyValue: Shapley Value Regression for Relative Importance of Attributes}
\description{
Shapley Value Regression for calculating the relative importance of independent variables in linear regression with avoiding the collinearity.
}
\keyword{internal}
