% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapleyvalue_parallel.R
\name{shapleyvalue_parallel}
\alias{shapleyvalue_parallel}
\title{Shapley Value Regression (parallel)}
\usage{
shapleyvalue_parallel(y, x)
}
\arguments{
\item{y}{is the data set of dependent variable}

\item{x}{is the data set of independent variables}
}
\description{
Shapley Value Regression (parallel)
}
