% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapleyValue.R
\name{shapleyvalue}
\alias{shapleyvalue}
\title{Shapley Value Regression}
\usage{
shapleyvalue(y, x)
}
\arguments{
\item{y}{is the data set of dependent variable}

\item{x}{is the data set of independent variables}
}
\description{
Shapley Value Regression
}
