\name{ShapleyValue}
\alias{ShapleyValue}
\title{ShapleyValueRegression -- to calculate the relative importance of attributes in linear regression}
\usage{
shapleyvalue(y = the dependent variable, x = independent variables)
}
\description{
Prints a data table with two rows: the unstandardized and standardized relative importance of each attributes using shapley value regression method.
}
\examples{
shapleyvalue(y,x)
}
